requirements.txt
beautifulsoup4==4.13.4
lxml==5.2.2
pandas==1.5.3
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.16.0
urllib3==2.2.2
# lyricsgenius and python-docx are optional.  They are required for network lyric fetching
# and document generation respectively.  Install them if you intend to use those features.
lyricsgenius==3.0.1
python-docx==1.1.2